move file with FlexibleAttributeForPDUService
